+++ src/salamatrix-sdk/native-ui-runtime/salamatrix_ui_win32_host.h
@@ -4,4 +4,5 @@
 
 namespace Salamatrix { namespace UI {
 INativeDialogHost* WINAPI GetWin32NativeDialogHost();
+void WINAPI SetWin32NativeDialogDarkMode(BOOL enabled);
 } }